decimal to float for transaction amout

+++ packages/montanabay39/mpesa/src/database/migrations/2021_05_09_000000_create_mpesa_transactions_table.php
@@ -19,7 +19,7 @@ class CreateMpesaTransactionsTable extends Migration
             $table->string('partyA'); // depositor. ['c2b / lnmo' => $MSISDN, 'b2c' => $ShortCode]
             $table->string('partyB'); // receiver. ['c2b / lnmo' => $ShortCode, 'b2c' => $MSISDN]
             $table->decimal('transactionType'); // transaction action category. ['in' => ['c2b', 'lnmo'], 'out' => 'b2c']
-            $table->decimal('transactionAmount'); // $Amount
+            $table->float('transactionAmount', 32); // $Amount
             $table->string('transactionCode')->nullable()->unique(); // Nullable on transaction initialization and filled in callback event.
             $table->timestamp('transactionTimeStamp'); // a transaction initialization time stamp.
             $table->json('transactionDetails'); // transaction remark's. e.t.c
